long -> int

+++ src/main/java/app/onepass/organizer/repositories/EventDurationRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import app.onepass.organizer.entities.EventDurationEntity;
 
-public interface EventDurationRepository extends JpaRepository<EventDurationEntity, Long> {
+public interface EventDurationRepository extends JpaRepository<EventDurationEntity, Integer> {
 
 	long deleteAllByEventId(long eventId);
 }
